refactor: Split timeseries and trip libs
